Site/requirements.txt
Flask
psycopg2-binary
Werkzeug
# Ajoute ici d'autres dépendances communes si besoin
